+++ cmd/l1cmd/validators.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2022, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package blockchaincmd
+package l1cmd
 
 import (
 	"errors"
