+++ cmd/nodecmd/status.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/ava-labs/avalanche-cli/cmd/blockchaincmd"
+	"github.com/ava-labs/avalanche-cli/cmd/l1cmd"
 	"github.com/ava-labs/avalanche-cli/pkg/ansible"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
 	"github.com/ava-labs/avalanche-cli/pkg/models"
@@ -78,7 +78,7 @@ func statusNode(_ *cobra.Command, args []string) error {
 	}
 	if blockchainName != "" {
 		// check subnet first
-		if _, err := blockchaincmd.ValidateSubnetNameAndGetChains([]string{blockchainName}); err != nil {
+		if _, err := l1cmd.ValidateSubnetNameAndGetChains([]string{blockchainName}); err != nil {
 			return err
 		}
 	}
