+++ cmd/transactioncmd/transaction_commit.go
@@ -5,7 +5,7 @@ package transactioncmd
 import (
 	"fmt"
 
-	"github.com/ava-labs/avalanche-cli/cmd/blockchaincmd"
+	"github.com/ava-labs/avalanche-cli/cmd/l1cmd"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
 	"github.com/ava-labs/avalanche-cli/pkg/keychain"
 	"github.com/ava-labs/avalanche-cli/pkg/subnet"
@@ -63,7 +63,7 @@ func commitTx(_ *cobra.Command, args []string) error {
 		return err
 	}
 	if !isPermissioned {
-		return blockchaincmd.ErrNotPermissionedSubnet
+		return l1cmd.ErrNotPermissionedSubnet
 	}
 	subnetAuthKeys, remainingSubnetAuthKeys, err := txutils.GetRemainingSigners(tx, controlKeys)
 	if err != nil {
@@ -73,7 +73,7 @@ func commitTx(_ *cobra.Command, args []string) error {
 	if len(remainingSubnetAuthKeys) != 0 {
 		signedCount := len(subnetAuthKeys) - len(remainingSubnetAuthKeys)
 		ux.Logger.PrintToUser("%d of %d required signatures have been signed.", signedCount, len(subnetAuthKeys))
-		blockchaincmd.PrintRemainingToSignMsg(subnetName, remainingSubnetAuthKeys, inputTxPath)
+		l1cmd.PrintRemainingToSignMsg(subnetName, remainingSubnetAuthKeys, inputTxPath)
 		return fmt.Errorf("tx is not fully signed")
 	}
 
@@ -94,7 +94,7 @@ func commitTx(_ *cobra.Command, args []string) error {
 
 	if txutils.IsCreateChainTx(tx) {
 		// TODO: teleporter for multisig
-		if err := blockchaincmd.PrintDeployResults(subnetName, subnetID, txID); err != nil {
+		if err := l1cmd.PrintDeployResults(subnetName, subnetID, txID); err != nil {
 			return err
 		}
 		return app.UpdateSidecarNetworks(&sc, network, subnetID, txID, "", "", sc.Networks[network.Name()].BootstrapValidators)
